almost done meshing togthere arduino comm and motor driver

+++ sim/src/ackdrive_arduino/src/arduino_comms.cpp
@@ -1,4 +1,4 @@
-#include "diffdrive/arduinoComms.h"
+#include "ackdrive_arduino/arduino_comms.h"
 // #include <ros/console.h>
 // #include <rclcpp/rclcpp.hpp> error here
 #include <sstream>
@@ -35,7 +35,7 @@ void ArduinoComms::readEncoderValues(int &val_1, int &val_2)
     val_2 = std::atoi(token_2.c_str());
 }
 
-void ArduinoComms::setMotorValues(int val_1, int val_2)
+void ArduinoComms::setMotorValues(double val_1, double val_2)
 {
     std::stringstream ss;
     ss << "m " << val_1 << " " << val_2 << "\r";
